style updates, new routes
