+++ frontend/src/routes/(app)/(home)/+layout.svelte
@@ -1,33 +1,43 @@
 <script lang="ts">
-	import { goto } from '$app/navigation';
 	import { page } from '$app/stores';
+	import Banner from '$lib/components/general/Banner.svelte';
+	import HomeHeader from '$lib/components/general/HomeHeader.svelte';
 	import { Tooltip } from '@svelte-plugins/tooltips';
 
 	export let data;
+
+	let isExplore =
+		$page.route.id === '/(app)/(home)/reports' || $page.route.id === '/(app)/(home)/projects';
 </script>
 
 <div class="flex flex-col w-full h-full p-6 pt-5 bg-yellowish">
-	{#if data.cognitoUser !== null}
+	{#if data.cognitoUser && data.cognitoUser !== null}
 		<div class="flex text-3xl">
 			<Tooltip content={'Your projects and reports'} theme={'zeno-tooltip'} position="bottom">
-				<button
-					class={`mr-6 ${
-						!$page.url.href.includes('my') ? 'text-grey-dark' : ''
-					} hover:text-primary`}
-					on:click={() => goto('/my')}
+				<a
+					href={'/' + data.cognitoUser.name + '/projects'}
+					class={`mr-6 ${isExplore ? 'text-grey-dark' : ''} hover:text-primary`}
 				>
 					My Hub
-				</button>
+				</a>
 			</Tooltip>
 			<Tooltip content={'Public projects and reports'} theme={'zeno-tooltip'} position="bottom">
-				<button
-					class={`${!$page.url.href.includes('public') ? 'text-grey-dark' : ''} hover:text-primary`}
-					on:click={() => goto('/public')}
-				>
+				<a href="/projects" class={`${!isExplore ? 'text-grey-dark' : ''} hover:text-primary`}>
 					Discover
-				</button>
+				</a>
 			</Tooltip>
 		</div>
+	{:else}
+		<Banner
+			>Welcome to <a class="text-primary font-medium" href="https://zenoml.com">Zeno</a>!
+			<a class="text-primary font-medium" href="/login">Sign in </a>
+			or <a class="text-primary font-medium" href="/signup">sign up</a> to create and see your projects
+			and reports.</Banner
+		>
 	{/if}
-	<slot />
+	<div class="p-4 mt-5 flex flex-col bg-white shadow">
+		<HomeHeader user={isExplore ? '' : data.cognitoUser?.name} />
+
+		<slot />
+	</div>
 </div>
