+++ frontend/src/routes/(app)/(home)/[user]/+layout.server.ts
