+++ frontend/src/routes/(app)/(home)/[user]/+page.server.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export async function load() {
+	throw redirect(303, './reports');
+}
