+++ frontend/src/routes/(app)/(home)/[user]/reports/+page.server.ts
@@ -1,10 +1,9 @@
 import { checkRefreshCookie } from '$lib/util/userCookieRefresh.js';
 import { getEndpoint } from '$lib/util/util';
-import { OpenAPI, ZenoService, type Project, type Report } from '$lib/zenoapi';
+import { OpenAPI, ZenoService, type Report } from '$lib/zenoapi';
 import { redirect } from '@sveltejs/kit';
 
 export async function load({ cookies, depends, url }) {
-	depends('app:projects');
 	depends('app:reports');
 
 	OpenAPI.BASE = `${getEndpoint()}/api`;
@@ -12,7 +11,6 @@ export async function load({ cookies, depends, url }) {
 	const userCookie = cookies.get('loggedIn');
 
 	let reports: Report[] = [];
-	let projects: Project[] = [];
 	if (!userCookie) {
 		throw redirect(303, '/');
 	}
@@ -22,10 +20,8 @@ export async function load({ cookies, depends, url }) {
 		Authorization: 'Bearer ' + cognitoUser.accessToken
 	};
 	reports = await ZenoService.getReports();
-	projects = await ZenoService.getProjects();
 
 	return {
-		projects: projects,
 		reports: reports
 	};
 }
